+++ commandline/src/test/java/tech/pegasys/web3signer/commandline/valueprovider/YamlConfigFileDefaultProviderTest.java
@@ -21,11 +21,13 @@ import static tech.pegasys.web3signer.CmdlineHelpers.validBaseCommandOptions;
 import tech.pegasys.web3signer.CmdlineHelpers;
 import tech.pegasys.web3signer.commandline.Web3SignerBaseCommand;
 import tech.pegasys.web3signer.commandline.valueprovider.DemoCommand.SubCommand;
+import tech.pegasys.web3signer.common.Web3SignerMetricCategory;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Set;
 
 import org.apache.logging.log4j.Level;
 import org.junit.jupiter.api.Test;
@@ -45,19 +47,27 @@ class YamlConfigFileDefaultProviderTest {
     commandLine.setDefaultValueProvider(new YamlConfigFileDefaultProvider(commandLine, configFile));
 
     final String cmdArgs =
-        removeFieldFrom(validBaseCommandOptions(), "http-listen-port", "key-store-path");
+        removeFieldFrom(validBaseCommandOptions(), "http-listen-port", "key-config-path");
     final String[] args = cmdArgs.split(" ");
     commandLine.parseArgs(args);
 
     assertThat(web3SignerBaseCommand.getHttpListenPort()).isEqualTo(6001);
     assertThat(web3SignerBaseCommand.getKeyConfigPath()).isEqualTo(Path.of("./keys_yaml"));
+    assertThat(web3SignerBaseCommand.getMetricCategories())
+        .isEqualTo(Set.of(Web3SignerMetricCategory.HTTP));
+    assertThat(web3SignerBaseCommand.getLogLevel()).isEqualTo(Level.INFO);
   }
 
   @Test
   void valuesFromConfigFileWithSubcommandArePopulated(@TempDir final Path tempDir)
       throws IOException {
     final String subcommandOptions =
-        String.join(lineSeparator(), "demo.x: 10", "demo.y: 2", "demo.name: \"test name\"");
+        String.join(
+            lineSeparator(),
+            "demo.x: 10",
+            "demo.y: 2",
+            "demo.name: \"test name\"",
+            "demo.aliases: \"test alias\"");
     final String config = CmdlineHelpers.validBaseYamlOptions() + subcommandOptions;
     final File configFile = Files.writeString(tempDir.resolve("config.yaml"), config).toFile();
 
@@ -74,6 +84,27 @@ class YamlConfigFileDefaultProviderTest {
     assertThat(subCommand.x).isEqualTo(10);
     assertThat(subCommand.y).isEqualTo(2);
     assertThat(subCommand.name).isEqualTo("test name");
+    assertThat(subCommand.alias).isEqualTo("test alias");
+  }
+
+  @Test
+  void aliasesFromConfigFileArePopulated(@TempDir final Path tempDir) throws IOException {
+    final String subcommandOptions = String.join(lineSeparator(), "demo.alias: \"test alias\"");
+    final String config = CmdlineHelpers.validBaseYamlAliasOptions() + subcommandOptions;
+    final File configFile = Files.writeString(tempDir.resolve("config.yaml"), config).toFile();
+    final Web3SignerBaseCommand web3SignerBaseCommand = new Web3SignerBaseCommand();
+    final DemoCommand subCommand = new DemoCommand();
+    final CommandLine commandLine = new CommandLine(web3SignerBaseCommand);
+    commandLine.registerConverter(Level.class, Level::valueOf);
+    commandLine.addSubcommand("demo", subCommand);
+    commandLine.setDefaultValueProvider(new YamlConfigFileDefaultProvider(commandLine, configFile));
+    commandLine.parseArgs("demo");
+
+    assertThat(web3SignerBaseCommand.getMetricCategories())
+        .isEqualTo(Set.of(Web3SignerMetricCategory.HTTP));
+    assertThat(web3SignerBaseCommand.getLogLevel()).isEqualTo(Level.INFO);
+    assertThat(web3SignerBaseCommand.getKeyConfigPath()).isEqualTo(Path.of("./keys_yaml_alias"));
+    assertThat(subCommand.alias).isEqualTo("test alias");
   }
 
   @Test